pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mdsa-framework"
version = "1.0.0"
description = "Multi-Domain Small Language Model Agentic Orchestration Framework"
readme = "README.md"
license = {text = "Apache 2.0"}
authors = [
    {name = "MDSA Team", email = "mdsa@example.com"}
]
keywords = ["ai", "machine-learning", "nlp", "orchestration", "slm", "small-language-models", "rag", "multi-domain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.2",
    "accelerate>=0.24.0",
    "peft>=0.6.0",
    "chromadb>=0.4.18",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=11.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "cryptography>=41.0.0",
    "psutil>=5.9.0",
    "pydantic>=2.5.0",
    "prometheus-client>=0.17.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]
gpu = [
    "bitsandbytes>=0.41.0",
]
faiss = [
    "faiss-cpu>=1.7.4",
]
faiss-gpu = [
    "faiss-gpu>=1.7.4",
]
pinecone = [
    "pinecone-client>=2.2.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/your-org/mdsa-framework"
Documentation = "https://mdsa-framework.readthedocs.io"
Repository = "https://github.com/your-org/mdsa-framework"
"Bug Tracker" = "https://github.com/your-org/mdsa-framework/issues"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | \.eggs
  | \.mypy_cache
  | \.pytest_cache
  | \.claude
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=mdsa",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "performance: marks tests as performance benchmarks",
]

# mypy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "sentence_transformers.*",
    "chromadb.*",
    "torch.*",
    "peft.*",
]
ignore_missing_imports = true

# Coverage configuration
[tool.coverage.run]
source = ["mdsa"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "@abc.abstractmethod",
]
